chnaged int to double

+++ 2022-RapidReact/src/main/java/frc/robot/Constants.java
@@ -85,6 +85,6 @@ public class Constants {
     }
 
     public static final class PressureConstants {
-        public static final int pressureValue = 80;
+        public static final double pressureValue = 80;
     }
 }
